Split auth related config for Azure
